+++ test/Main.hs
@@ -10,8 +10,10 @@ import CryptoTest
 main :: IO ()
 main = do
     passphraseCbTests <- CryptoTest.cbTests
+    keyCbTests <- KeyTest.cbTests
     defaultMain $ testGroup "tests"
         [ testGroup "key"    KeyTest.tests
+        , keyCbTests
         , testGroup "keyGen" KeyGenTest.tests
         , testGroup "ctx"    CtxTest.tests
         , CryptoTest.tests